csv-cleaner/requirements.txt
pandas>=2.2.0
pyarrow>=15.0.0
openpyxl>=3.1.2
typer[all]>=0.12.3
pydantic>=2.6.0
pyyaml>=6.0.1